requirements.txt
minio==7.1.14
apache-airflow-providers-docker==3.7.1
apache-airflow-providers-postgres
apache-airflow-providers-slack==8.1.0